+++ docs/storybook/src/frontstage/Modal.stories.tsx
@@ -19,6 +19,9 @@ const meta = {
     },
     layout: "fullscreen",
   },
+  args: {
+    notifyCloseRequest: false,
+  },
 } satisfies Meta<typeof ModalFrontstageStory>;
 
 export default meta;
@@ -37,5 +40,17 @@ export const BackButton: Story = {
         }}
       />
     ),
+  },
+};
+
+export const AppBarRight: Story = {
+  args: {
+    appBarRight: <>appBarRight</>,
+  },
+};
+
+export const NotifyCloseRequest: Story = {
+  args: {
+    notifyCloseRequest: true,
   },
 };