+++ ui/core-react/CHANGELOG.md
@@ -1,5 +1,17 @@
 # Change Log - @itwin/core-react
 
+## 5.17.0
+
+### Minor Changes
+
+- 2e99cd6: Added `offset` property to `ContextMenu` component.
+
+## 5.16.0
+
+### Minor Changes
+
+- 51fb78f: Added support for [React v19](https://react.dev/blog/2024/12/05/react-19).
+
 ## 5.15.2
 
 ## 5.15.1